Don't close STDOUT when running test code
This causes failures on perl 5.24, and is a bad idea in general.
Since file descriptor 1 (STDOUT) is closed in the child process,
a subsequent call to open() may populate file descriptor 1 with
the new file descriptor, which may or may not even be open for writing.
If it's not, calls to write() will fail, but if it *is*, seemingly
innocent calls that write to standard output (such as those in the
standard C library) can clobber the contents of a seemingly unrelated
file, which is *not* fun to track down.