Treat openat as open in tests
Later versions of glibc (2.26 and beyond - see
https://lwn.net/Articles/738694/) use openat() rather than
open() when talking to the OS kernel.  Since my tests are
looking for a call to open(), this breaks my tests!

The solution (for the tests, anyway) is to screen for both open and
openat, and to transform the openat calls to their equivalent open
calls before comparison.  I can't just convert to openat wholesale,
because then the tests would break for glibc pre-2.26